Refacttor de calculador de puntso de usuario
