--- src/main/java/ar/edu/unq/desapp/grupoo022020/backenddesappapi/model/IPointSystemState.java
@@ -1,7 +1,0 @@
-package ar.edu.unq.desapp.grupoo022020.backenddesappapi.model;
-
-public interface IPointSystemState {
-
-	public void givePointsToUser(Donation donation);
-
-}
