+++ src/main/java/ar/edu/unq/desapp/grupoo022020/backenddesappapi/service/IPointSystemState.java
@@ -1,0 +1,9 @@
+package ar.edu.unq.desapp.grupoo022020.backenddesappapi.service;
+
+import ar.edu.unq.desapp.grupoo022020.backenddesappapi.model.Donation;
+
+public interface IPointSystemState {
+
+	public void givePointsToUser(Donation donation);
+
+}
