+++ src/main/java/ar/edu/unq/desapp/grupoo022020/backenddesappapi/service/InvestInProyectWithMoreThanTwoThousandHabitants.java
@@ -1,4 +1,7 @@
-package ar.edu.unq.desapp.grupoo022020.backenddesappapi.model;
+package ar.edu.unq.desapp.grupoo022020.backenddesappapi.service;
+
+import ar.edu.unq.desapp.grupoo022020.backenddesappapi.model.Donation;
+import ar.edu.unq.desapp.grupoo022020.backenddesappapi.model.UserDonator;
 
 public class InvestInProyectWithMoreThanTwoThousandHabitants implements IPointSystemState {
