+++ src/main/java/ar/edu/unq/desapp/grupoo022020/backenddesappapi/service/PointCalculatorContext.java
@@ -1,10 +1,12 @@
-package ar.edu.unq.desapp.grupoo022020.backenddesappapi.model;
+package ar.edu.unq.desapp.grupoo022020.backenddesappapi.service;
+
+import ar.edu.unq.desapp.grupoo022020.backenddesappapi.model.Donation;
 
 final class PointCalculatorContext {
 
-	private static IPointSystemState getPointSystem(Donation donation) {
+	private static IPointSystemState getPointSystem(Donation donation, Integer donationInMonth) {
 		IPointSystemState system = null;
-		if (donation.getDonationInMonth() > 0) {
+		if (donationInMonth > 0) {
 			return new InvestInMoreThanOneProjectInCalendarMonth();
 		}
 		if (donation.getInvestment() >= 1000.00F) {
@@ -16,8 +18,8 @@ final class PointCalculatorContext {
 		return system;
 	}
 
-	public static void givePointsToUser(Donation donation) {
-		IPointSystemState system = getPointSystem(donation);
+	public static void givePointsToUser(Donation donation, Integer donationInMonth) {
+		IPointSystemState system = getPointSystem(donation, donationInMonth);
 		if (system != null) {
 			system.givePointsToUser(donation);
 		}
